refactor(routers): move admin read routes
