+++ docs/context/architecture/super-admin.md
@@ -3,6 +3,7 @@ title: Super-admin — cross-tenant read + control
 status: shipped
 sources:
   - src/treg/api.py
+  - src/treg/routers/admin.py
   - src/treg/routers/dependencies.py
   - src/treg/config.py
 related:
@@ -27,6 +28,10 @@ flag (so a web portal can log in with either). Returns a principal string (for a
 
 `api.py` re-exports this dependency during the staged router migration, so the existing admin route
 definitions and compatibility imports still refer to the same function object.
+
+The cross-tenant GET and reconciliation handlers live in `routers.admin`; `api.py` appends its two
+router blocks around the unchanged admin mutation block to preserve registration order. The mutation
+handlers and their control helpers remain in `api.py` for the later control-layer extraction.
 
 ## Suspension enforcement (in `require_member`)
 Two flags gate the **org-scoped** path: `require_member` raises 403 if `user.suspended` ("account
